AP_Logger: support 32 bit system IDs
Widen source/target system IDs in the MAVC log message and the
MAVLink log transfer target.

+++ libraries/AP_Logger/AP_Logger.h
@@ -278,7 +278,7 @@ public:
 
     void Write_EntireMission();
     void Write_Command(const mavlink_command_int_t &packet,
-                       uint8_t source_system,
+                       uint32_t source_system,
                        uint8_t source_component,
                        MAV_RESULT result,
                        bool was_command_long=false);
